create common header comp
